Time: 4 ms (38.42%), Space: 65.2 MB (67.3%) - LeetHub

+++ 0713-subarray-product-less-than-k/0713-subarray-product-less-than-k.cpp
@@ -3,7 +3,7 @@ public:
     int numSubarrayProductLessThanK(vector<int>& nums, int k) {
         int n = nums.size();
         int l = 0, ans = 0;
-        long long prod = 1;
+        int prod = 1;
         for (int r = 0; r < n; r++) {
             prod *= nums[r];
             while (l <= r && prod >= k) {
